nlp-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
nltk==3.8.1
scikit-learn==1.3.2
python-dotenv==1.0.0